Use dependency update instead of build
It turned out that `dependency build` requires the repos of
dependencies to be added to the local config whereas
`dependency update` does not. We, thus, have to change the
logic here. An update always regenerates the lock file
which we don't really want as it is expected to be up to
date. However, adding repos is not easily possible. We'd
have to parse that information out of Chart.yaml/requirement.yaml.

Signed-off-by: Reinhard Nägele <unguiculus@gmail.com>