Update releaser to regenerate generated field on index update
Currently, the only time the `generated` field is ever written is when
we invoke the repo.NewIndexFile() field. This means that contrary to
what you would expect, the `generated` field is not updated when the
index is updated.

This adds a single line change that calls `time.Now()` after we have
determined that the index is to be updated, and updates the `generated`
field before the index is written.

Signed-off-by: Nikolas Skoufis <nskoufis@seek.com.au>